% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_digit_empty.R
\name{predict_digit_empty}
\alias{predict_digit_empty}
\title{Predicts the digit based on a gray-scale image}
\usage{
predict_digit_empty(image)
}
\arguments{
\item{image}{a 784 (28x28) length vector with gray-scale values between 0-255.}
}
\value{
the predicted digit which is always 0
}
\description{
Predicts the digit based on a gray-scale image
}
